requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
PyJWT==2.10.1
cryptography==43.0.3
httpx==0.28.0
python-dotenv==1.0.1
uuid6
